# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Credit Broker 2.0",
        "description": "Документация\n * Во всех ответах приходит объект: `data`, кроме ошибок: В ошибках `401` имеется только поле `message`; в ошибке `422` имеются поля: `errors и message`.\n * Для авторизованных запросов необходимо вставить в HEADER `Authorization ` Bearer токен полученный при авторизации. Пример:\n *     Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n * Токен можно получить при входе в систему или регистрации. Сам токен будет расположен в ответе HEADER:\n *      X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n *      X-Access-Token-Expires-At: 2023-11-08 10:13:18\n *  Важно! На всех запросах, кроме DEV - необходимо отправлять в HEADER Platform-Code и Device-Id. Иначе будет приходить  `403 ошибка`\n *      Platform-Code: ios\n *      Device-Id: DFIE3232dD\n *\n * На всех запросах в HEADER ОБЯЗАТЕЛЬНО установить:\n *     Accept: application/json",
        "contact": {
            "email": "develop@tsay.uz"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API LOCAL SERVER"
        },
        {
            "url": "http://{tenant}.localhost",
            "description": "API LOCAL SERVER WITH TENANT",
            "variables": {
                "tenant": {
...
            }
...
    "paths": {
...
        "/v1/profile/auth/code": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Запрос на получение OTP кода",
                "operationId": "980792548c543d23b400fdfc40148859",
                "requestBody": {
                    "description": "Тело запроса OTP аутентификации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "phone": "+998909101828"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/v1/profile/auth/verify": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Запрос на верификацию OTP кода",
                "operationId": "a26e62d6a45f89a5b79adb95119687ac",
                "requestBody": {
                    "description": "Тело запроса OTP верификации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "phone": "+998909101828",
                                "otp": "789456"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/v1/profile/auth/logout": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Запрос на выход из системы",
                "operationId": "1719dce0e27ce363b87912d2681e4ee5",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "v1.Profile.UserResource": {
                "description": "Class UserResource.",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer",
                        "example": ""
                    },
                    "phone": {
                        "description": "Номер телефона пользователя",
                        "type": "string",
                        "example": ""
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": ""
                    },
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": ""
                    },
                    "is_active": {
                        "description": "Статус пользователя",
                        "type": "boolean",
                        "example": ""
                    },
                    "created_at": {
                        "description": "Дата создания сотрудника",
                        "type": "string",
                        "format": "date-time",
                        "example": ""
                    },
                    "updated_at": {
                        "description": "Дата обновления сотрудника",
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Пример: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjI0OT",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "Platform-Code": {
                "type": "apiKey",
                "description": "Код платформы",
                "name": "Platform-Code",
                "in": "header"
            },
            "Device-Id": {
                "type": "apiKey",
                "description": "ID устройства",
                "name": "Device-Id",
                "in": "header"
            },
            "Company-slug": {
                "type": "apiKey",
                "description": "Slug компании",
                "name": "Company",
                "in": "header"
            }
        }
    }
}